Move all the stories
