+++ packages/pwa/src/components/atoms/Button.stories.tsx
@@ -1,5 +1,5 @@
 import {Button} from '@src/components/atoms/Button';
-import {StorySection} from '@src/components/styleguide/StorySection';
+import {StorySection} from '@src/components/stories/StorySection';
 
 import {toast} from '@src/lib/toasts';
