+++ packages/pwa/src/components/atoms/Divider.stories.tsx
@@ -1,6 +1,6 @@
 import {Divider} from '@src/components/atoms/Divider';
 import {Text} from '@src/components/atoms/Text';
-import {StorySection} from '@src/components/styleguide/StorySection';
+import {StorySection} from '@src/components/stories/StorySection';
 
 export const DividerStories: React.FC = () => {
   return (
