+++ packages/pwa/src/components/atoms/Flex.stories.tsx
@@ -1,6 +1,6 @@
 import {FlexColumn, FlexRow} from '@src/components/atoms/Flex';
 import {Text} from '@src/components/atoms/Text';
-import {StorySection} from '@src/components/styleguide/StorySection';
+import {StorySection} from '@src/components/stories/StorySection';
 
 import {vars} from '@src/lib/theme.css';
