+++ packages/pwa/src/components/atoms/Link.stories.tsx
@@ -1,6 +1,6 @@
 import {Link} from '@src/components/atoms/Link';
 import {Text} from '@src/components/atoms/Text';
-import {StorySection} from '@src/components/styleguide/StorySection';
+import {StorySection} from '@src/components/stories/StorySection';
 
 export const LinkStories: React.FC = () => {
   return (
