+++ packages/pwa/src/components/atoms/Spacer.stories.tsx
@@ -1,6 +1,6 @@
 import {Spacer} from '@src/components/atoms/Spacer';
 import {Text} from '@src/components/atoms/Text';
-import {StorySection} from '@src/components/styleguide/StorySection';
+import {StorySection} from '@src/components/stories/StorySection';
 
 export const SpacerStories: React.FC = () => {
   return (
